rework module structure
