+++ cloudshell/cp/core/flows/__init__.py
@@ -1,0 +1,4 @@
+from .cleanup_sandbox_infra import AbstractCleanupSandboxInfraFlow  # noqa
+from .deploy import AbstractDeployFlow  # noqa
+from .prepare_sandbox_infra import AbstractPrepareSandboxInfraFlow  # noqa
+from .vm_details import AbstractVMDetailsFlow  # noqa
