+++ cloudshell/cp/core/request_actions/cleanup_sandbox_infra.py
@@ -1,0 +1,27 @@
+import json
+from dataclasses import dataclass
+
+from cloudshell.cp.core.requests import models
+from cloudshell.cp.core.requests.actions.base import BaseRequestActions
+
+
+@dataclass
+class CleanupSandboxInfraRequestActions(BaseRequestActions):
+    cleanup_network: models.CleanupNetwork = None
+
+    @classmethod
+    def from_request(cls, request, cs_api=None):
+        """
+
+        :param request:
+        :param cs_api:
+        :return:
+        """
+        actions = cls._parse_request_actions(request=request, cs_api=cs_api)
+        obj = cls()
+
+        for action in actions:
+            if isinstance(action, models.CleanupNetwork):
+                obj.cleanup_network = action
+
+        return obj
