+++ cloudshell/cp/core/request_actions/models/connectivity.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import Any
 
-from cloudshell.cp.core.models.base import BaseRequestAction, BaseRequestObject
+from cloudshell.cp.core.requests.models import BaseRequestAction, BaseRequestObject
 
 
 @dataclass
