+++ cloudshell/cp/core/request_actions/models/deploy_app.py
@@ -1,7 +1,7 @@
 import itertools
 from dataclasses import dataclass, field
 
-from cloudshell.cp.core.models.base import BaseRequestAction, BaseRequestObject
+from cloudshell.cp.core.requests.models import BaseRequestAction, BaseRequestObject
 
 
 @dataclass
