--- tests/cloudshell/cp/core/test_request_actions.py
@@ -1,11 +1,0 @@
-import unittest
-
-from cloudshell.cp.core.request_actions import PrepareSandboxInfraRequestActions
-
-
-class TestPrepareSandboxInfraRequestActions(unittest.TestCase):
-    def setUp(self):
-        self.request_actions = PrepareSandboxInfraRequestActions
-
-    def test_resource_model(self):
-        pass
